adding components as per feedback: herosection, OTPVerificationModal, VirtualCard
